refactor: 💡 Rename *.mcss files to *.module.css
This has several benefits:
* It's accepted as a best practice by the CSS module specification community:
  https://github.com/css-modules/css-modules/issues/229
* It's the accepted standard for Vue CLI:
  https://cli.vuejs.org/guide/css.html#css-modules
* It's the accepted standard for Gatsby:
  https://www.gatsbyjs.org/packages/gatsby-plugin-react-css-modules/?=module
* It preserves a pure ".css" extension for files, allowing more IDEs and tools
  to recognize the file format.

BREAKING CHANGE: Webpack loader rules must be updated to match the new file pattern.
