+++ Documentation/config.js
@@ -39,7 +39,7 @@ module.exports = {
         ],
       },
       {
-        test: /\\.mcss$/,
+        test: /\\.module\.css$/,
         use: [
           { loader: 'style-loader' },
           {
