+++ Sources/Interaction/UI/CornerAnnotation/CornerAnnotation.module.css
