+++ Sources/Interaction/UI/FPSMonitor/FPSMonitor.module.css
