+++ Utilities/config/rules-vtk.js
@@ -15,7 +15,7 @@ module.exports = [
     ],
   },
   {
-    test: /\.mcss$/,
+    test: /\.module\.css$/,
     use: [
       { loader: 'style-loader' },
       {
